requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
neo4j
redis
python-dotenv
beautifulsoup4
google-generativeai
google-genai
openai